Sencha adapter "filter" key can't use "id" property #202
